Reword users

+++ functions/src/functions/users.ts
@@ -10,30 +10,20 @@ import admin from 'firebase-admin'
 import { https } from 'firebase-functions'
 import { type CallableRequest, onCall } from 'firebase-functions/v2/https'
 import { type Result } from './types.js'
-import {
-  type UserAuth,
-  type Admin,
-  type Clinician,
-  type Patient,
-  type User,
-} from '../models/user.js'
+import { type UserAuth, type User, UserType } from '../models/user.js'
+import { Credential, UserRole } from '../services/credential.js'
 import { CacheDatabaseService } from '../services/database/cacheDatabaseService.js'
 import { FirestoreService } from '../services/database/firestoreService.js'
-import { SecurityService } from '../services/securityService.js'
 import { DatabaseUserService } from '../services/user/databaseUserService.js'
 import { type UserService } from '../services/user/userService.js'
 
 export interface GetUsersInformationInput {
-  includeClinicianData?: boolean
-  includePatientData?: boolean
   includeUserData?: boolean
   userIds?: string[]
 }
 
 export interface UserInformation {
   auth: UserAuth
-  clinician?: Clinician
-  patient?: Patient
   user?: User
 }
 
@@ -58,19 +48,19 @@ export const getUsersInformationFunction = onCall(
       throw new https.HttpsError('not-found', 'User not found')
     const organization = authenticatedUser.content.organization
 
-    const securityService = new SecurityService()
-    await securityService.ensureClinician(request.auth, organization)
+    const credential = new Credential(request.auth, userService)
+    await credential.checkAny(
+      UserRole.admin,
+      ...(organization ? [UserRole.clinician(organization)] : []),
+    )
 
     const result: GetUsersInformationOutput = {}
     for (const userId of request.data.userIds) {
       try {
         const userData = await userService.getUser(userId)
-        // organization is undefined for admins
         if (organization && userData?.content.organization !== organization)
-          throw new https.HttpsError(
-            'permission-denied',
-            'User does not belong to the same organization',
-          )
+          throw credential.permissionDeniedError()
+
         const user = await userService.getAuth(userId)
         const userInformation: UserInformation = {
           auth: {
@@ -79,14 +69,6 @@ export const getUsersInformationFunction = onCall(
             phoneNumber: user.phoneNumber,
             photoURL: user.photoURL,
           },
-        }
-        if (request.data.includeClinicianData ?? false) {
-          const clinician = await userService.getClinician(userId)
-          userInformation.clinician = clinician?.content
-        }
-        if (request.data.includePatientData ?? false) {
-          const patient = await userService.getPatient(userId)
-          userInformation.patient = patient?.content
         }
         if (request.data.includeUserData ?? false) {
           userInformation.user = userData?.content
@@ -144,12 +126,11 @@ export const updateUserInformationFunction = onCall(
     if (!user) throw new https.HttpsError('not-found', 'User not found')
     const organization = user.content.organization
 
-    const securityService = new SecurityService()
-    try {
-      await securityService.ensureClinician(request.auth, organization)
-    } catch {
-      securityService.ensureUser(request.auth, request.data.userId)
-    }
+    const credential = new Credential(request.auth, userService)
+    await credential.checkAny(
+      ...(organization ? [UserRole.clinician(organization)] : []),
+      UserRole.user(request.data.userId),
+    )
 
     await userService.updateAuth(request.data.userId, request.data.data.auth)
   },
@@ -171,13 +152,14 @@ export const deleteUserFunction = onCall(
       new CacheDatabaseService(new FirestoreService()),
     )
     const user = await userService.getUser(request.auth.uid)
+    if (!user) throw new https.HttpsError('not-found', 'User not found')
 
-    const securityService = new SecurityService()
-    await securityService.ensureClinician(
-      request.auth,
-      user?.content.organization,
+    const credential = new Credential(request.auth, userService)
+    await credential.checkAny(
+      ...(user.content.organization ?
+        [UserRole.clinician(user.content.organization)]
+      : []),
     )
-
     await userService.deleteUser(request.data.userId)
     return 'Success'
   },
@@ -185,9 +167,6 @@ export const deleteUserFunction = onCall(
 
 export interface CreateInvitationInput {
   auth?: UserAuth
-  admin?: Admin
-  clinician?: Clinician
-  patient?: Patient
   user?: User
 }
 
@@ -209,19 +188,25 @@ export const createInvitationFunction = onCall(
     if (!request.data.user)
       throw new https.HttpsError('invalid-argument', 'User data is required')
 
-    const securityService = new SecurityService()
-    if (request.data.admin !== undefined)
-      await securityService.ensureAdmin(request.auth)
-    else
-      await securityService.ensureOwner(
-        request.auth,
-        request.data.user.organization,
-      )
+    const userService: UserService = new DatabaseUserService(
+      new CacheDatabaseService(new FirestoreService()),
+    )
+    const credential = new Credential(request.auth, userService)
+    if (request.data.user.type === UserType.admin) {
+      await credential.checkAny(UserRole.admin)
+    } else if (request.data.user.organization) {
+      await credential.checkAny(UserRole.owner(request.data.user.organization))
+    } else {
+      throw credential.permissionDeniedError()
+    }
 
     const firestore = admin.firestore()
     const invitationCollection = firestore.collection('invitations')
     const invitationCode =
-      request.data.clinician !== undefined || request.data.admin !== undefined ?
+      (
+        request.data.user.type === UserType.admin ||
+        request.data.user.type === UserType.clinician
+      ) ?
         request.data.auth.email
       : undefined
 
@@ -254,79 +239,3 @@ function generateInvitationCode(length: number): string {
   }
   return result
 }
-
-export interface GrantOwnerInput {
-  userId?: string
-  organizationId?: string
-}
-
-export const grantOwnerFunction = onCall(
-  async (request: CallableRequest<GrantOwnerInput>) => {
-    if (!request.data.userId)
-      throw new https.HttpsError('invalid-argument', 'User ID is required')
-    if (!request.data.organizationId)
-      throw new https.HttpsError(
-        'invalid-argument',
-        'Organization ID is required',
-      )
-    const securityService = new SecurityService()
-    await securityService.grantOwner(
-      request.auth,
-      request.data.userId,
-      request.data.organizationId,
-    )
-    return 'Success'
-  },
-)
-
-export interface RevokeOwnerInput {
-  userId?: string
-  organizationId?: string
-}
-
-export const revokeOwnerFunction = onCall(
-  async (request: CallableRequest<RevokeOwnerInput>) => {
-    if (!request.data.userId)
-      throw new https.HttpsError('invalid-argument', 'User ID is required')
-    if (!request.data.organizationId)
-      throw new https.HttpsError(
-        'invalid-argument',
-        'Organization ID is required',
-      )
-    const securityService = new SecurityService()
-    await securityService.revokeOwner(
-      request.auth,
-      request.data.userId,
-      request.data.organizationId,
-    )
-    return 'Success'
-  },
-)
-
-export interface GrantAdminInput {
-  userId?: string
-}
-
-export const grantAdminFunction = onCall(
-  async (request: CallableRequest<GrantAdminInput>) => {
-    const securityService = new SecurityService()
-    if (!request.data.userId)
-      throw new https.HttpsError('invalid-argument', 'User ID is required')
-    await securityService.grantAdmin(request.auth, request.data.userId)
-    return 'Success'
-  },
-)
-
-export interface RevokeAdminInput {
-  userId?: string
-}
-
-export const revokeAdminFunction = onCall(
-  async (request: CallableRequest<RevokeAdminInput>) => {
-    if (!request.data.userId)
-      throw new https.HttpsError('invalid-argument', 'User ID is required')
-    const securityService = new SecurityService()
-    await securityService.revokeAdmin(request.auth, request.data.userId)
-    return 'Success'
-  },
-)
